Two tests that fail in the MCS kernel
This adds two tests, one which causes an actual failure in the kernel
and one which indicates a precondition on a function which isn't
obvious.

The first test unbinds an SC from a thread blocked on recv (although the
same could be done for a thread blocked on send with a passive server
receiving and achieve much the same result). This leads to an assertion
failure when the kernel switches to a thread with a bound SC that is not
ready to run.

The second case is handled by the master kernel but includes a call in
the kernel to `refill_unblock_check` where the scheduling context is not
configured. This case is handled by the `isRoundRobin` branch as an
unconfigured SC satisfies the check for a round robin SC.